Now able to add a spell through the website. The interaction between the api code and the website code seems to be periodically sending an empty body followed by the proper body. When this happens, the initial error is EOF. For now, I am having the handler verify that err == nil before attempting further processing. It is working for now. I will do more tomorrow.

+++ app/models/school.go
@@ -13,7 +13,7 @@ import (
 )
 
 type School struct {
-	SchoolID    int    `json:"school_id"`
+	SchoolID    int64  `json:"school_id"`
 	Name        string `json:"name"`
 	Summary     string `json:"summary"`
 	Description string `json:"description"`
@@ -25,13 +25,13 @@ type School struct {
 }
 
 type SchoolSummary struct {
-	SchoolID int    `json:"school_id"`
+	SchoolID int64  `json:"school_id"`
 	Name     string `json:"name"`
 	Summary  string `json:"summary"`
 }
 
 type SchoolHeader struct {
-	SchoolID int    `json:"school_id"`
+	SchoolID int64  `json:"school_id"`
 	Name     string `json:"name"`
 }
